requirements.txt
streamlit
huggingface-hub
openai
neo4j
langchain_groq
python-dotenv
crewai
pandas
langchain
langchain_community
langchain_openai